Separated `gen/type.go` info multiple files
